+++ player/js/animation/AnimationItem.js
@@ -22,6 +22,7 @@ var AnimationItem = function () {
     this.renderer = null;
     this.animationID = randomString(10);
     this.scaleMode = 'fit';
+    this.assetsPath = '';
     this.timeCompleted = 0;
     this.segmentPos = 0;
     this.segments = [];
@@ -80,6 +81,7 @@ AnimationItem.prototype.setParams = function(params) {
         }else{
             this.path = params.path.substr(0,params.path.lastIndexOf('/')+1);
         }
+        this.assetsPath = params.assetsPath;
         this.fileName = params.path.substr(params.path.lastIndexOf('/')+1);
         this.fileName = this.fileName.substr(0,this.fileName.lastIndexOf('.json'));
         xhr.open('GET', params.path, true);
@@ -165,7 +167,6 @@ AnimationItem.prototype.includeLayers = function(data) {
     //this.animationData.tf = 50;
     dataManager.completeData(this.animationData,this.renderer.globalData.fontManager);
     this.renderer.includeLayers(data.layers);
-    this.renderer.buildStage(this.container, this.layers);
     if(expressionsPlugin){
         expressionsPlugin.initExpressions(this);
     }
@@ -230,6 +231,10 @@ AnimationItem.prototype.configAnimation = function (animData) {
     this.firstFrame = Math.round(this.animationData.ip);
     this.frameMult = this.animationData.fr / 1000;
     this.trigger('config_ready');
+    this.imagePreloader = new ImagePreloader();
+    this.imagePreloader.setAssetsPath(this.assetsPath);
+    this.imagePreloader.setPath(this.path);
+    this.imagePreloader.loadAssets(animData.assets);
     this.loadSegments();
     this.updaFrameModifier();
     if(this.renderer.globalData.fontManager){
@@ -267,7 +272,6 @@ AnimationItem.prototype.elementLoaded = function () {
 
 AnimationItem.prototype.checkLoaded = function () {
     if (this.pendingElements === 0) {
-        this.renderer.buildStage(this.container, this.layers);
         if(expressionsPlugin){
             expressionsPlugin.initExpressions(this);
         }
@@ -541,6 +545,22 @@ AnimationItem.prototype.updaFrameModifier = function () {
 
 AnimationItem.prototype.getPath = function () {
     return this.path;
+};
+
+AnimationItem.prototype.getAssetsPath = function (assetData) {
+    var path = '';
+    if(this.assetsPath){
+        var imagePath = assetData.p;
+        if(imagePath.indexOf('images/') !== -1){
+            imagePath = imagePath.split('/')[1];
+        }
+        path = this.assetsPath + imagePath;
+    } else {
+        path = this.path;
+        path += assetData.u ? assetData.u : '';
+        path += assetData.p;
+    }
+    return path;
 };
 
 AnimationItem.prototype.getAssetData = function (id) {