+++ player/js/elements/BaseElement.js
@@ -14,6 +14,12 @@ BaseElement.prototype.checkMasks = function(){
     return false;
 }
 
+BaseElement.prototype.checkParenting = function(){
+    if(this.data.parent !== undefined){
+        this.comp.buildElementParenting(this, this.data.parent);
+    }
+}
+
 BaseElement.prototype.prepareFrame = function(num){
     if(this.data.ip - this.data.st <= num && this.data.op - this.data.st > num)
     {
@@ -57,6 +63,20 @@ BaseElement.prototype.prepareFrame = function(num){
     this.currentFrameNum = num*this.data.sr;
     return this.isVisible;
 };
+
+BaseElement.prototype.initExpressions = function(){
+    this.layerInterface = LayerExpressionInterface(this);
+    //layers[i].layerInterface = LayerExpressionInterface(layers[i]);
+    if(this.data.hasMask){
+        this.layerInterface.registerMaskInterface(this.maskManager);
+    }
+    if(this.data.ty === 0){
+        this.compInterface = CompExpressionInterface(this);
+    } else if(this.data.ty === 4){
+        this.layerInterface.shapeInterface = ShapeExpressionInterface.createShapeInterface(this.shapesData,this.viewData,this.layerInterface);
+    }
+}
+
 BaseElement.prototype.setBlendMode = function(){
     var blendModeValue = '';
     switch(this.data.bm){