+++ player/js/elements/CompElement.js
@@ -2,6 +2,7 @@ function ICompElement(data,parentContainer,globalData,comp, placeholder){
     this._parent.constructor.call(this,data,parentContainer,globalData,comp, placeholder);
     this.layers = data.layers;
     this.isSvg = true;
+    this.elements = Array.apply(null,{length:this.layers.length});
     if(this.data.tm){
         this.tm = PropertyFactory.getProp(this,this.data.tm,0,globalData.frameRate,this.dynamicProperties);
     }
@@ -16,7 +17,9 @@ ICompElement.prototype.hide = function(){
     if(!this.hidden){
         var i,len = this.elements.length;
         for( i = 0; i < len; i+=1 ){
-            this.elements[i].hide();
+            if(this.elements[i]){
+                this.elements[i].hide();
+            }
         }
         this.hidden = true;
     }
@@ -37,7 +40,12 @@ ICompElement.prototype.prepareFrame = function(num){
     this.renderedFrame = timeRemapped/this.data.sr;
     var i,len = this.elements.length;
     for( i = 0; i < len; i+=1 ){
-        this.elements[i].prepareFrame(timeRemapped/this.data.sr - this.layers[i].st);
+        if(!this.elements[i]){
+            this.checkLayer(i, this.renderedFrame - this.layers[i].st, this.layerElement);
+        }
+        if(this.elements[i]){
+            this.elements[i].prepareFrame(timeRemapped/this.data.sr - this.layers[i].st);
+        }
     }
 };
 
@@ -51,10 +59,8 @@ ICompElement.prototype.renderFrame = function(parentMatrix){
 
     this.hidden = false;
     for( i = 0; i < len; i+=1 ){
-        if(this.data.hasMask){
+        if(this.elements[i]){
             this.elements[i].renderFrame();
-        }else{
-            this.elements[i].renderFrame(this.finalTransform);
         }
     }
     if(this.firstFrame){
@@ -77,3 +83,18 @@ ICompElement.prototype.destroy = function(){
         this.elements[i].destroy();
     }
 };
+
+ICompElement.prototype.checkLayer = BaseRenderer.prototype.checkLayer;
+ICompElement.prototype.buildItem = BaseRenderer.prototype.buildItem;
+ICompElement.prototype.createItem = SVGRenderer.prototype.createItem;
+ICompElement.prototype.buildElementParenting = SVGRenderer.prototype.buildElementParenting;
+ICompElement.prototype.createItem = SVGRenderer.prototype.createItem;
+ICompElement.prototype.createImage = SVGRenderer.prototype.createImage;
+ICompElement.prototype.createComp = SVGRenderer.prototype.createComp;
+ICompElement.prototype.createSolid = SVGRenderer.prototype.createSolid;
+ICompElement.prototype.createShape = SVGRenderer.prototype.createShape;
+ICompElement.prototype.createText = SVGRenderer.prototype.createText;
+ICompElement.prototype.createPlaceHolder = SVGRenderer.prototype.createPlaceHolder;
+ICompElement.prototype.createBase = SVGRenderer.prototype.createBase;
+ICompElement.prototype.buildItemParenting = SVGRenderer.prototype.buildItemParenting;
+ICompElement.prototype.appendElementInPos = SVGRenderer.prototype.appendElementInPos;