+++ player/js/elements/canvasElements/CVImageElement.js
@@ -1,6 +1,5 @@
 function CVImageElement(data, comp,globalData){
     this.assetData = globalData.getAssetData(data.refId);
-    this.path = globalData.getPath();
     this._parent.constructor.call(this,data, comp,globalData);
     this.globalData.addPendingElement();
 }
@@ -39,7 +38,8 @@ CVImageElement.prototype.createElements = function(){
     this.img = new Image();
     this.img.addEventListener('load', imageLoaded, false);
     this.img.addEventListener('error', imageFailed, false);
-    this.img.src = this.path+this.assetData.p;
+    var assetPath = this.globalData.getAssetsPath(this.assetData);
+    this.img.src = assetPath;
 
     this._parent.createElements.call(this);